+++ runtime/runtime_impl_jetty12/src/main/java/com/google/apphosting/runtime/jetty/JettyServletEngineAdapter.java
@@ -23,14 +23,13 @@ import com.google.apphosting.base.protos.RuntimePb.UPRequest;
 import com.google.apphosting.base.protos.RuntimePb.UPResponse;
 import com.google.apphosting.runtime.AppEngineConstants;
 import com.google.apphosting.runtime.AppVersion;
-import com.google.apphosting.runtime.AppEngineConstants;
+import com.google.apphosting.runtime.LocalRpcContext;
 import com.google.apphosting.runtime.MutableUpResponse;
 import com.google.apphosting.runtime.ServletEngineAdapter;
 import com.google.apphosting.runtime.anyrpc.EvaluationRuntimeServerInterface;
 import com.google.apphosting.runtime.jetty.delegate.DelegateConnector;
 import com.google.apphosting.runtime.jetty.delegate.impl.DelegateRpcExchange;
 import com.google.apphosting.runtime.jetty.http.JettyHttpHandler;
-import com.google.apphosting.runtime.jetty.http.LocalRpcContext;
 import com.google.apphosting.runtime.jetty.proxy.JettyHttpProxy;
 import com.google.apphosting.utils.config.AppEngineConfigException;
 import com.google.apphosting.utils.config.AppYaml;
