+++ runtime/runtime_impl_jetty12/src/main/java/com/google/apphosting/runtime/jetty/http/JettyHttpHandler.java
@@ -25,16 +25,14 @@ import com.google.apphosting.runtime.ApiProxyImpl;
 import com.google.apphosting.runtime.AppVersion;
 import com.google.apphosting.runtime.BackgroundRequestCoordinator;
 import com.google.apphosting.runtime.AppEngineConstants;
+import com.google.apphosting.runtime.LocalRpcContext;
 import com.google.apphosting.runtime.RequestManager;
 import com.google.apphosting.runtime.RequestRunner;
 import com.google.apphosting.runtime.RequestRunner.EagerRunner;
 import com.google.apphosting.runtime.ResponseAPIData;
 import com.google.apphosting.runtime.ServletEngineAdapter;
-import com.google.apphosting.runtime.AppEngineConstants;
 import com.google.apphosting.runtime.jetty.AppInfoFactory;
-import com.google.common.base.Ascii;
 import com.google.common.flogger.GoogleLogger;
-import org.eclipse.jetty.http.HttpField;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.Response;
@@ -45,7 +43,6 @@ import org.eclipse.jetty.util.Callback;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.time.Duration;
-import java.util.Optional;
 import java.util.concurrent.TimeoutException;
 
 import static com.google.apphosting.runtime.RequestRunner.WAIT_FOR_USER_RUNNABLE_DEADLINE;
